refactor: 사진 권한 확인을 repo로 분리
